requirements.txt
# AUTOGENERATED, DO NOT EDIT

asgiref==3.7.2
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==5.0.1
django-admin-ordering==0.17.0
django-bootstrap3==23.6
django-canonical-domain==0.9.2
django-debug-toolbar==4.3.0
django-js-asset==2.2.0
django-vanilla-views==3.0.0
flake8==7.0.0
gunicorn==21.2.0
httplib2==0.22.0
icalendar==5.0.11
mccabe==0.7.0
oauth2client==4.1.3
packaging==23.2
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pyflakes==3.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
rsa==4.9
setuptools==69.0.3
six==1.16.0
speckenv==6.2
sqlparse==0.4.4
typing_extensions==4.9.0
wheel==0.42.0
